# docs/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders/checkout": {
      "post": {
        "operationId": "OrdersController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/pay": {
      "post": {
        "operationId": "OrdersController_confirmPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders": {
      "get": {
        "operationId": "OrdersController_findMyOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Boogie Neko API",
    "description": "The official API for the Boogie Neko store",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "shippingAddressId": {
            "type": "number",
            "description": "The Id of address where the products are shipped to"
          }
        },
        "required": [
          "shippingAddressId"
        ]
      },
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "number",
            "description": "The id of order that this payment belongs to"
          },
          "transactionId": {
            "type": "string",
            "description": "Transaction ID of the payment for the order",
            "example": "112233445566"
          }
        },
        "required": [
          "orderId",
          "transactionId"
        ]
      },
...
    }
  }
}